logging/log.h: move enum class formatter to a separate file ...
... to common/logging/formatter.h
