+++ src/common/CMakeLists.txt
@@ -85,6 +85,7 @@ add_library(common STATIC
     logging/backend.h
     logging/filter.cpp
     logging/filter.h
+    logging/formatter.h
     logging/log.h
     logging/log_entry.h
     logging/text_formatter.cpp
