+++ src/common/logging/log.h
@@ -6,25 +6,11 @@
 
 #include <algorithm>
 #include <string_view>
-#include <type_traits>
 
 #include <fmt/format.h>
 
+#include "common/logging/formatter.h"
 #include "common/logging/types.h"
-
-// adapted from https://github.com/fmtlib/fmt/issues/2704
-// a generic formatter for enum classes (<= 32 bits)
-#if FMT_VERSION >= 80100
-template <typename T>
-struct fmt::formatter<T, std::enable_if_t<std::is_enum_v<T>, char>>
-    : formatter<std::underlying_type_t<T>> {
-    template <typename FormatContext>
-    auto format(const T& value, FormatContext& ctx) -> decltype(ctx.out()) {
-        return fmt::formatter<std::underlying_type_t<T>>::format(
-            static_cast<std::underlying_type_t<T>>(value), ctx);
-    }
-};
-#endif
 
 namespace Common::Log {
